+++ packages/hooks/src/internal/useApiClient.ts
@@ -32,12 +32,14 @@ export function useApiClient(options: { optional?: boolean } = {}): ApiClient | 
     return new ApiClient({
       appKey: context.appKey,
       apiHost: context.apiHost,
+      timeout: context.timeout,
       installationId: context.installationId,
       additionalHeaders: context.additionalHeaders,
     });
   }, [
     context?.appKey,
     context?.apiHost,
+    context?.timeout,
     context?.installationId,
     context?.additionalHeaders,
     optional,